# Conflicts:
#	kantek/plugins/private/stats.py
#	kantek/plugins/private/update.py
#	kantek/plugins/private/user.py
#	kantek/utils/_config.py
#	kantek/utils/client.py

+++ kantek/plugins/private/update.py
@@ -1,36 +1,39 @@
-import asyncio
 from asyncio import subprocess
 
-from utils._config import Config
+from utils import helpers
 from utils.client import Client
 from utils.mdtex import *
 from utils.pluginmgr import k, Command
+from utils.tags import Tags
 
 
 @k.command('update')
-async def update(client: Client, event: Command) -> None:
+async def update(client: Client, event: Command, tags: Tags) -> None:
     """Run git pull and exit.
 
     This command assumes the bot is running under a process manager that automatically restarts it.
 
+    Tags:
+        update: Set to `silent` to silence any messages
+
     Examples:
         {cmd}
     """
-    config = Config()
-    progess_message = await client.respond(event, MDTeXDocument(
-        Section('Updating',
-                f'Running {Code("git pull")}')))
-    await subprocess.create_subprocess_shell('git pull -q')
-    await asyncio.sleep(3)
-    proc = await subprocess.create_subprocess_shell('git rev-parse --short HEAD',
-                                                    stdout=subprocess.PIPE,
-                                                    stderr=subprocess.PIPE)
-    new_commit = (await proc.stdout.read()).decode().strip()
-    await progess_message.delete()
-    await client.respond(
-        event,
-        MDTeXDocument(
-            Section('Updated Kantek',
-                    KeyValueItem('New commit', Link(new_commit, f'{config.source_url}/commit/{new_commit}')),
-                    Italic('Restarting bot'))))
+    silent = tags.get('update', False)
+    if not silent:
+        progess_message = await client.respond(event, MDTeXDocument(
+            Section('Updating',
+                    f'Running {Code("git pull")}')))
+    else:
+        await event.delete()
+    subprocess.call(['git', 'pull', '-q'])
+    new_commit = helpers.get_commit()
+    if not silent:
+        await progess_message.delete()
+        await client.respond(
+            event,
+            MDTeXDocument(
+                Section('Updated Kantek',
+                        KeyValueItem('New commit', Link(new_commit, helpers.link_commit(new_commit))),
+                        Italic('Restarting bot'))))
     await client.disconnect()
